refactor(accumulator): use shared gateway Network; fold version helper
Address review feedback on #482.

- Network: add a shared `Network` enum to `templar-gateway-client`
  (clap::ValueEnum behind a `clap` feature) and make the accumulator its
  first adopter, removing the bot's duplicate definition. This is the
  off-chain home for the type; remaining tools/services (and the
  templar-common copy) migrate incrementally in follow-up work.
- Helpers: fold `market_version` into a single private
  `supports_static_yield`, dropping a one-caller indirection and the
  public surface. Behaviour (skip markets < 1.1.0, conservative on
  undeterminable version) is unchanged and still mirrors harvest.
