+++ service/accumulator/Cargo.toml
@@ -16,7 +16,7 @@ futures = { workspace = true }
 near-account-id = { workspace = true }
 near-api = { workspace = true }
 templar-common = { workspace = true, features = ["rpc"] }
-templar-gateway-client = { workspace = true }
+templar-gateway-client = { workspace = true, features = ["clap"] }
 templar-gateway-methods-spec = { workspace = true }
 templar-gateway-types = { workspace = true }
 tokio = { workspace = true }
